requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
PyGithub>=2.1.1
